backend/pyproject.toml
[project]
name = "cop-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.26",
    "flask>=3.1.1",
    "numpy>=2.2.6",
    "psycopg-binary>=3.2.9",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
]
